Extract `Searcher` interface
By moving the number of threads from the old `parSearch` call to the
constructor for `ThreadedSearch` we can make both `LinearSearch` and
`ThreadedSearch` implement the same `Searcher` interface, which is nice.

While doing this I also replaced all the overly specific references to
`ArrayList` to just `List`, and replaced the old-style `for` loop in
`LinearSearch` with a `foreach` style loop.

I also added some nifty new timing computations to `Main` so we know
how long things take to complete.

Closes #3